Refactor: arraySize int형에서 size_t 형으로 변경

+++ cpp07/ex01/incs/iter.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   iter.hpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jihoh <jihoh@student.42seoul.kr>           +#+  +:+       +#+        */
+/*   By: jihoh <jihoh@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 17:44:13 by jihoh             #+#    #+#             */
-/*   Updated: 2022/06/21 17:49:51 by jihoh            ###   ########.fr       */
+/*   Updated: 2022/07/02 13:40:16 by jihoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,9 @@
 # define ITER_CLASS_H
 
 template <typename T>
-
-void iter(T *array, int arraySize, void (*f)(const T &))
+void iter(T *array, size_t arraySize, void (*f)(const T &))
 {
-	for (int i = 0; i < arraySize; i++)
+	for (size_t i = 0; i < arraySize; i++)
 		f(array[i]);
 }
